platform/backend/requirements.txt
# Deep Sci-Fi Platform Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.5  # Vector similarity search

# AI/Embeddings
openai>=1.0.0

# Voice Agent (AG-UI protocol)
pydantic-ai-slim[ag-ui,anthropic]>=1.50.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Security
slowapi>=0.1.9  # Rate limiting

# Observability (optional - requires LOGFIRE_TOKEN to activate)
logfire[fastapi,sqlalchemy,asyncpg,httpx,openai,system-metrics]>=3.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiohttp>=3.9.0
hypothesis>=6.100.0
